--- Client/src/Views/Landing/ContactLanding/ContactLanding.module.css
@@ -1,87 +1,0 @@
- .contactpagewrapper h1 {
-    max-width: 900px !important;
-  }
-  .contactpagewrapper {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    margin: 6rem 0rem;
-  }
-
-  .primaryheading {
-    font-size: clamp(2rem, 5vw, 4rem);
-    color: #4c4c4c;
-    max-width: 600px;
-  }
-
-
-
-  .contactformcontainer {
-    background-color: white;
-    max-width: 700px;
-    width: 100%;
-    margin-top: 3rem;
-    display: flex;
-    align-items: center;
-    padding: 1rem;
-    border-radius: 5rem;
-  }
-  .contactformcontainer input {
-    flex: 1;
-    height: 100%;
-    border: none;
-    outline: none;
-    font-size: 1.3rem;
-    padding: 0.5rem 1rem;
-  }
-  .secondarybutton {
-    padding: 1rem 2.5rem;
-    background-color: #fe9e0d;
-    outline: none;
-    border: none;
-    border-radius: 5rem;
-    font-size: 1.1rem;
-    cursor: pointer;
-    font-weight: 600;
-    color: white;
-    transition: 0.2s;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  .secondarybutton svg {
-    margin-left: 0.75rem;
-    font-size: 1.5rem;
-  }
-  .secondarybutton:hover {
-    background-color: #e48f0f;
-  }
-
-  @media (max-width: 800px){
-    .primaryheading {
-        text-align: center;
-        max-width: 90%;
-      }
-      .contact-form-container {
-        padding: 0.5rem !important;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        border-radius: 1rem !important;
-      }
-      .contact-form-container input {
-        font-size: 1rem !important;
-        max-width: 100%;
-        padding: 0.5rem 1rem !important;
-        margin-bottom: 0.8rem !important;
-        text-align: center;
-      }
-      .secondarybutton {
-        font-size: 1rem;
-        padding: 0.8rem 2rem;
-        cursor: pointer;
-      }
-    
-
-  }
