--- Client/src/Views/Landing/HomeLanding/HomeLanding.module.css
@@ -1,95 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: "Reem Kufi", sans-serif !important;
-  }
-
-.homebannercontainer {
-    position: relative;
-    display: flex;
-    padding-top: 3rem;
-  }
-  .primaryheading {
-    font-size: clamp(2rem, 5vw, 4rem);
-    color: #4c4c4c;
-    /* line-height: 5rem; */
-    max-width: 600px;
-  }
-  .primarytext {
-    font-size: clamp(1rem, 3vw, 1.5rem);
-    max-width: 500px;
-    color: #6a6a6a;
-    margin: 1.5rem 0rem;
-  }
-  .homeimagesection {
-    max-width: 600px;
-    flex: 1;
-  }
-
-  .secondarybutton {
-    padding: 1rem 2.5rem;
-    background-color: #fe9e0d;
-    outline: none;
-    border: none;
-    border-radius: 5rem;
-    font-size: 1.1rem;
-    cursor: pointer;
-    font-weight: 600;
-    color: white;
-    transition: 0.2s;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  .secondarybutton svg {
-    margin-left: 0.75rem;
-    font-size: 1.5rem;
-  }
-  .secondarybutton:hover {
-    background-color: #e48f0f;
-  }
-
-  @media (max-width: 800px){
-    .homebannercontainer,
-    .aboutsectioncontainer {
-      flex-direction: column-reverse;
-      align-items: center;
-      justify-content: center;
-    }
-    .secondarybutton {
-        font-size: 1rem;
-        padding: 0.8rem 2rem;
-        cursor: pointer;
-      }
-
-      .homeimagesection,
-      .aboutsectionimagecontainer {
-        width: 100%;
-        max-width: 400px;
-      }
-  
-  .hometextsection {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
-
-  }
-    .hometextsection,
-    .aboutsectiontextcontainer {
-      justify-content: center;
-      align-items: center;
-      margin-top: 4rem;
-    }
-
-    .primaryheading {
-        text-align: center;
-        max-width: 90%;
-      }
-      .primarytext {
-        text-align: center;
-        max-width: 80%;
-      }
-    }
