--- Client/src/Views/Landing/NavbarLanding/NavbarLanding.module.css
@@ -1,12 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: "Reem Kufi", sans-serif !important;
-  }
-
-  @media (max-width: 800px) {
-    .navlogocontainer {
-      max-width: 140px;
-    }
-}
