--- Client/src/Views/Landing/WorkLanding/WorkLanding.module.css
@@ -1,88 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: "Reem Kufi", sans-serif !important;
-  }
-
-  .worksectionwrapper {
-    margin-top: 15rem;
-  }
-
-  .worksectiontop p {
-    text-align: center;
-    max-width: 600px !important;
-  }
-  .worksectiontop h1 {
-    max-width: 700px !important;
-  }
-  .worksectiontop {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-  }
-
-  .primaryheading {
-    font-size: clamp(2rem, 5vw, 4rem);
-    color: #4c4c4c;
-    /* line-height: 5rem; */
-    max-width: 600px;
-  }
-  .primarytext {
-    font-size: clamp(1rem, 3vw, 1.5rem);
-    max-width: 500px;
-    color: #6a6a6a;
-    margin: 1.5rem 0rem;
-  }
-
-  .worksectionbottom {
-    margin-top: 5rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-  }
-  .worksectioninfo {
-    width: 290px;
-    min-height: 350px;
-    background-color: white;
-    padding: 1rem 2rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    border-radius: 1rem;
-    color: #505050;
-    margin: 1rem 2rem;
-  }
-  .worksectioninfo h2 {
-    margin: 1rem 0rem;
-  }
-  .worksectioninfo p {
-    flex: 1;
-    display: flex;
-    align-items: center;
-    font-weight: 600;
-  }
-
-  @media (max-width: 800px) {
-    .worksectionwrapper {
-        margin-top: 5rem !important;
-      }
-
-      .primaryheading {
-        text-align: center;
-        max-width: 90%;
-      }
-
-      .primarytext {
-        text-align: center;
-        max-width: 80%;
-      }
-
-      .worksectionbottom {
-        margin-top: 1rem !important;
-      }
-  }
